add section header cmp
