+++ components/Help.jsx
@@ -1,18 +1,16 @@
 import { HELPSTEPS } from "@/lib/helpSteps";
 import { styled } from "@/stitches.config";
 import React from "react";
-import { Heading, HelpCard, Text } from ".";
+import { Heading, HelpCard, SectionHeader, Text } from ".";
 
 export const Help = () => {
   return (
     <Wrapper>
-      <Header>
-        <Heading>How do we help?</Heading>
-        <Text>
-          We help you unlock all your hidden potentials by taking on several
-          missions until you become the badass that you wanna be!!
-        </Text>
-      </Header>
+      <SectionHeader
+        title="How do we help?"
+        description="We help you unlock all your hidden potentials by taking on several
+          missions until you become the badass that you wanna be!!"
+      />
 
       <HelpCards>
         {HELPSTEPS.map((step) => (
@@ -25,14 +23,6 @@ export const Help = () => {
 
 const Wrapper = styled("section", {
   marginTop: 100,
-});
-
-const Header = styled("div", {
-  textAlign: "center",
-  marginBottom: 64,
-  display: "flex",
-  flexDirection: "column",
-  gap: 24,
 });
 
 const HelpCards = styled("div", {
